changes for ppcmac
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@3219 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/GC/ioa.c
@@ -400,7 +400,7 @@ void DoIOACell(struct Object **theCell,struct Object *theObj)
   }
 #ifdef RTDEBUG
   else {
-    if ((long)theObj!=CALLBACKMARK){
+    if (theObj!=CALLBACKMARK){
       fprintf(output, 
 	      "DoIOACell: 0x%x: 0x%x is outside BETA heaps!\n", theCell, theObj);
       Illegal();
